refactor where HTMLGrid exists
