+++ quadratic-client/src/gridGL/HTMLGrid/htmlCells/htmlCellsHandler.ts
