+++ quadratic-client/src/gridGL/HTMLGrid/multiplayerInput/MultiplayerCellEdits.tsx
@@ -1,15 +1,8 @@
 import { MultiplayerCellEdit } from './MultiplayerCellEdit';
 import { useMultiplayerCellEdit } from './useMultiplayerCellEdit';
 
-interface Props {
-  container?: HTMLDivElement;
-}
-
-export const MultiplayerCellEdits = (props: Props) => {
-  const { container } = props;
+export const MultiplayerCellEdits = () => {
   const multiplayerCellInput = useMultiplayerCellEdit();
-
-  if (!container) return;
 
   return (
     <div style={{ pointerEvents: 'none' }}>
